22.04.22 11:18

+++ src/main/java/ru/kata/spring/boot_security/demo/dao/UserDao.java
@@ -12,11 +12,11 @@ public interface UserDao {
 
     public void addUser(User user);
 
-    public User getUserById(int id);
+    public User getUserById(long id);
 
     public User updateUser(User user);
 
-    public void removeUserById(int id);
+    public void removeUserById(long id);
 
     public List<User> listUsers();
 }
