+++ src/main/java/ru/kata/spring/boot_security/demo/dao/UserDaoImpl.java
@@ -34,7 +34,7 @@ public class UserDaoImpl implements UserDao {
     }
 
     @Override
-    public User getUserById(int id) {
+    public User getUserById(long id) {
         return entityManager.find(User.class, id);
     }
 
@@ -44,7 +44,7 @@ public class UserDaoImpl implements UserDao {
     }
 
     @Override
-    public void removeUserById(int id) {
+    public void removeUserById(long id) {
         entityManager.remove(entityManager.find(User.class, id));
     }
